chore(benches): switch to the criterion benchmarking framework
