+++ Cargo.toml
@@ -9,13 +9,16 @@ keywords = [ "Safe", "Network", "SafeNetwork", "SafeUrl", "XorUrl" ]
 authors = [ "MaidSafe Developers <dev@maidsafe.net>" ]
 edition = "2018"
 
+[features]
+dkg = ["bls_dkg"]
+
 [dependencies]
 serde_json = "1.0.64"
 thiserror = "1.0.24"
 quickcheck = "1"
 quickcheck_macros = "1"
 rand = "0.7.1"
-bls_dkg = "~0.3.8"
+bls_dkg = { version = "~0.3.8", optional = true }
 threshold_crypto = "0.4"
 
   [dependencies.tiny-keccak]
@@ -30,3 +33,16 @@ threshold_crypto = "0.4"
   [dependencies.serde]
   version = "1.0.111"
   features = [ "derive", "rc" ]
+
+[dev-dependencies]
+sn_dbc = {path = ".", features = ["dkg"] } #  HACK: enable the dkg features in tests
+criterion = "0.3"
+
+
+[profile.test]
+require-features = ["dkg"]
+
+[[bench]]
+name = "reissue"
+harness = false
+required-features = ["dkg"]
